system_distributed_keyspace: Break out snapshot entry structs to header
To avoid cross-inclusion in dependents
